requirements.txt
pytest-testinfra
pytest-xdist
paramiko
